backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
pymongo==4.4.0
pandas==2.2.3
python-dotenv==1.0.0
sentence-transformers
scikit-learn==1.5.1   # (fallback if faiss fails)
langdetect==1.0.9     # optional, for multilingual support
